.github/workflows/requirements.txt
python-hcl2
atlassian-python-api
bs4
